portal 2 page data no longer appears in spanish (?? why was it in spanish), star rating now works, ratings can be decimal, added theme and lastSynced to user model, changed sync logic to not replace existing names, descriptions, or tags

+++ backend/src/main/java/com/gametracker/backend/dto/entry/EntryInfoDTO.java
@@ -16,10 +16,10 @@ public record EntryInfoDTO(
         Integer currentAchievements,
         String reviewText,
         LocalDate finishDate,
-        Integer rating,
-        Integer enjoyment,
-        Integer gameplay,
-        Integer story,
-        Integer visuals,
-        Integer sound
+        Double rating,
+        Double enjoyment,
+        Double gameplay,
+        Double story,
+        Double visuals,
+        Double sound
 ) {}
